refactor: 공통 클래스를 common 패키지로 하위로 이동
