+++ app/src/main/java/woowacourse/shopping/presentation/common/model/CartProductUiModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.presentation.model
+package woowacourse.shopping.presentation.common.model
 
 import woowacourse.shopping.domain.model.CartProduct
