+++ app/src/main/java/woowacourse/shopping/presentation/common/model/FetchPageDirection.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.presentation.model
+package woowacourse.shopping.presentation.common.model
 
 enum class FetchPageDirection {
     PREVIOUS,
