+++ app/src/main/java/woowacourse/shopping/presentation/view/order/suggestion/adapter/SuggestionProductDiffUtil.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.presentation.view.order.suggestion.adapter
 
 import androidx.recyclerview.widget.DiffUtil
-import woowacourse.shopping.presentation.model.SuggestionProductUiModel
+import woowacourse.shopping.presentation.common.model.SuggestionProductUiModel
 
 object SuggestionProductDiffUtil : DiffUtil.ItemCallback<SuggestionProductUiModel>() {
     override fun areItemsTheSame(
