+++ app/src/main/res/layout/item_cart.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="cartItem"
-            type="woowacourse.shopping.presentation.model.CartProductUiModel" />
+            type="woowacourse.shopping.presentation.common.model.CartProductUiModel" />
 
         <variable
             name="eventListener"
