+++ app/src/main/res/layout/item_product.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="product"
-            type="woowacourse.shopping.presentation.model.CatalogItem.ProductItem" />
+            type="woowacourse.shopping.presentation.common.model.CatalogItem.ProductItem" />
 
         <variable
             name="adapterPosition"
